requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-pinecone>=0.1.0
streamlit>=1.28.0
pypdf>=3.17.0
pinecone-client>=3.0.0
openai>=1.0.0
tiktoken>=0.5.0
chromadb>=0.4.0